split utils.py and rewrite training requirements file
